Mondrian Project - Placing The Grid Items
The rest grid items, except from the gap grid items, were placed on the layout.
Some changes were added to the HTML and CSS, by adding an "id" attribute to the
"div" elements of the following patter "grid-item-x", where "x" is the number of
the item in the custom design picture.